Modified oversampling factor to be a float with valid values between 1.0 and 2.0
Ruven indicated that sampling factors > 2.0 should theoretically yield no additional quality improvements.  This was empirically confirmed through testing of resize operations using PNG lossless output with embedded ICC color profiles and Lanczos filtering.

+++ src/Main.cc
@@ -290,9 +290,9 @@ int main( int argc, char *argv[] )
   // Get the allow upscaling setting
   bool allow_upscaling = Environment::getAllowUpscaling();
 
-  // Get the oversampling factor which will result in requesting a tile 
-  // N times larger than would otherwise be requested prior to resize operations
-  int oversamplingFactor = Environment::getOversamplingFactor();
+  // Get the oversampling factor f which will result in requesting a tile 
+  // at least f times larger than the requested output dimensions
+  float oversamplingFactor = Environment::getOversamplingFactor();
 
 #ifdef HAVE_PNG
   // Get PNG Compression Level
